Add red dot for debugging highest point to place flower roots
updateFlowerRootHighestPoint() once debugged should be able to find a fixed point on the teacher's prairie background. This can be used to set the highest point that the flower grid can place flowers, particularly the "roots" of the flower's. Once we think that it is accurate, we can remove the pointer and use the Y value given for flower grid placement.