Refine the flower grid fixed point method
Further refinement of our adaption of the solution by Daniel Imms found at @ https://stackoverflow.com/a/15838104 

Some testing and input may be needed because I have tested different screen sizes by eyeing the dot placement and it seems to work. However, some sort of unit or e2e testing could be used to make sure the math is airtight. 

The hard coded 83 pixel componentHeight offset (initially put in by us to offset the top navbar and whitespace displacing the image) doesn't seem to be needed with the way that the method works. I would think it's needed, but not as far as I can tell.