Implement the API changes specified in the previous commit
Both Month and Quarter are now indexed from one instead of zero, and Month is now an enum.

We're using one-based indices because they're familiar—months already have a well-entrenched numbering scheme starting from one, and it would be confusing to use zero-based indices in that context. (See the discussion on PR #29 for more details.)

This commit makes the Karma tests pass again.