Move the top menu bar out of AppComponent
In particular, this commit makes the top menu bar its own component.

This gives us the flexibility to decide whether we want a given page to have a top menu bar or not. (In particular, in the future, we probably want to get rid of the top menu bar on the large display page.