Change mdc-top-app-bar from `fixed` to `fixedAdjustElement`
`fixedAdjustElement` is nice because it automatically gives the contents of the page a vertical offset so that the menu bar doesn't cover them up.

(In order to use this functionality, though, we have to tell mdc-top-app-bar which DOM element to give the offset to, which means we have to add an @Input to TopMenuBarComponent. This isn't a huge deal, but we do have to change it in a few places.)

Previously, we were putting in this vertical offset by hand. This is less than optimal for a few reasons:

1. It was brittle, since we depended on the exact height of the menu bar in pixels, and

2. The offset had to be applied to every page in the project, which meant that we couldn't easily remove the menu bar from a page.

The second reason is the one I care about, since I want to remove the menu bar from the large display page.