Add a bottom-bar to the large display page
It has a bunch of buttons, but they aren't hooked up to anything yet.

I'd like to comment on a few implementation decisions.

First:

My thought is that the .large-display-content div will eventually have an ngSwitch directive on it; that way, you can display different content at during different parts of the round. (The .bottom-bar div will probably do something similar.)

Second:

I opted against using an mdcIconButton for the play-pause button. (Instead, I'm using a regular mdcButton.) I know that this is a bit of an abuse of mdcButtons—they're really meant to have text inside, not icons.

That being said, an mdcIconButton looks out of place here—it's too small, and it doesn't have the right rounded-rect shape. I opted for aesthetics over correctness.

Third:

The bottom-bar layout is a bit complicated; essentially, I'm breaking it into three subsections: left, right, and center. (That way, some elements can be left-justified, while others can be centered or right-justified at the same time.)

Fourth:

At some point, we're probably going  to put a little three-dots menu on the right side of the bottom-bar--I just haven't gotten around to that yet :)