Reposition the flowers
 So that they look good at a variety of aspect ratios.

At extremely very wide aspect ratios (wider than 16:9) some of the flowers are partially out of the frame, and the display becomes hard to use. That being said, this problem used to be much worse; the new layout greatly ameliorates things.

In positioning the flowers, I've tried to strike a balance between a clear, easy-to-read grid and a naturalistic staggered look. However, it may still need some tweaking in the future.