Make sure that Nest objects have an `id` field
Like Bee and Flower objects.

This id field is a unique identifier for that type of nest; it's the key by which that Nest can be accessed in the global `nests` dictionary.

Technically speaking, you can always just get a Nest's id by doing a reverse lookup in the `nests` dictionary. That being said, we thought it would be helpful to denormalize the data a bit so you don't have to go through all that trouble.