Test the data we're importing from JSON
These tests mostly make sure that the data is in the right format, and they especially test the processing we do on it.

In particular, we test the parts where we replace ISO date strings with TimePeriods, and we test the part where we replace flower ids with FlowerSpecies objects.

These tests all pass already.