Rename all sorts of things
We were having a really hard time figuring out how to name these variables. We want to make the distinction between individual flowers and flower species, but we were having trouble with the fact that the word "species" is both singular and plural.

Eventually, we settled on Paul's suggestion, which is to use "allFlowerSpecies" for the collection of species and "flowerSpecies" for an individual species.