Add ability to add user to Firebase Database
User id and display name are added to the firebase database. 

Note: There are currently still some issues with this implementation. For one, the user isn't added the first time their account is created. Additionally, whenever you update the user information such as the display name, the page must be refreshed in order to show the changes.

Addresses issue #54

Co-authored-by: Audrey Le Meur <lemeu001@umn.edu>