Factor out the "querying Firebase" logic
That logic used to live in the StudentSessionService and StudentRoundService. Now, it lives in a few methods in the FirebaseService class. You pass in the IDs of the firebase documents you want, and it gives you observables of the session or of the round.

Separating out this logic will make testing easier later; we have the flexibility to mock those methods on the FirebaseService.