Add a bit more superstructure
Sorry, I'll get to writing actual tests eventually ;-) But in the meantime, I wanted to make a few more changes to the testing environment.

First, `scheduledIt()` is a function that's useful to us in several places, so this commit moves it into its own file. (That way, if you want to use it, you can just import it.)

Secondly,  we gave the mockFirebaseService two mock sessions it can read from, so that we can test switching between them.