Improve the null checking in StudentRoundService.currentBeeSpecies$
This fixes the tests from the previous commit.

There was a problem where, *if* roundStudentData$ was non-null, *but* it was missing its `beeSpecies` attribute, then the student's bee species ID was being reported as `undefined` when it should have been reported as `null`. This was a mostly harmless bug, but it made distinctUntilChanged() behave very wonkily.