Create a component to show while the teacher waits to start a round
It's just a stub at the moment.

Eventually, this component will be shown on the large display page while the students are joining the round. It'll probably show the round code and the list of students who've currently joined. (But that's off in the future.)

Really, all we need right now is the "start the round" button, which will cause the round in Firebase to be populated with data.