Add exported test data for use in the Firebase emulator
This adds a small set of test data, "demo-session" and "demo-round" that matches what we have in production Firestore at the moment.
It can be used with the `--import` parameter when running the Firebase emulators. This also creates a npm script `npm run firebase:emulator` that starts the emulator for just Firestore and imports the test data.

Co-authored-by: KK Lamberty <lamberty@morris.umn.edu>