Create a stub version TeacherRoundService
At Paul's recommendation, we're going to transfer a lot of the functionality of LargeDisplayComponent into TeacherRoundService.

We're going to do test-driven-development here, so I've already written unit tests for TeacherRoundService. (Several of them are failing, of course.)

I haven't touched the LargeDisplayComponent yet, though—we'll edit that after we know that TeacherRoundService works.