Make sure that .debug is overlayed on the AR view correctly
It needs to have position: absolute, or else it pushes the rest of the divs on the screen around, and you get a weird white border at the bottom of the screen.

I'm afraid I don't know the exact mechanism by which the .debug div was messing things up; all I know is that the layout was off :( Sorry for not having more details!