Assign bees to leftover students
When the number of students don't perfectly match the weights of the bees, there might be some leftover. We randomly assign a bee to these students. (At the moment, they have an equal chance of being any of the available bees)

For example, if there are two types of bees with an equal chance of being assigned to a student and there are 5 students, 2 will be assigned to one bee type and 2 to the other. There will be one "leftover" student who is then randomly assigned to one of the two bee types.

Co-authored-by: Audrey Le Meur <lemeu001@umn.edu>