Fix the StudentSessionService tests
We moved some methods around in commit aeb7de1, and the specs needed to be updated accordingly.

Viz:
The method that registers the student as a member of the session used to be named "AuthService.addStudentToDatabase()"; now it's "StudentSessionService.joinSession()".

Additionally, the method that tells the student-side which session to read from right now used to be called "StudentSessionService.joinSession()"; now it's called "StudentSessionService.setCurrentSession()".

Apologies for the confusion!

(The Karma tests as a whole are still failing, but that's just because we haven't written the specs for StudentDisplayComponent yet. We'll get there.)