Write the Contributing, Code of Conduct, and Acknowledgements sections
The Contributing section is mostly just a link to the newly-created `DEVELOPMENT.md` file, which doesn't have much information in it yet. (Most of the rest of this branch will be spent writing `DEVELOPMENT.md`.) I copied the Code of Conduct and Acknowledgements sections from the README in Pollinator Facts.