Make some slight adjustments to the GitHub Packages section
First, we changed the heading from a gerund phrase to an imperative sentence ("Telling" -> "Tell"). This is just a personal preference.

Second, we removed a line explaining how to open the command line. Now that we've factored out the "Getting Started" material, the audience of this document probably knows how to get to the command line.