Update `Interaction`s in the firestore tests
In this branch, Interactions got the `isNest` field, so when we're
making Interactions as test data, we need to change them all so that
they each have an `isNest`.

That makes all of the tests pass! (I haven't actually compiled and run
the website, though, so I can't guarrantee that it works right now.)