Add a back button to the about page
It's not the prettiest design in the world, but it's clear and obvious—which is important for navigation—and it uses the same design language as the rest of our website. Still, though, if any of you have ideas for how to make the back button better, I'd greatly appreciate that—I'm not entirely satisfied with it.