Add some more tests to the host page
Testing the 'reconnect to session' button, to make sure that it only appears at the right times. (In the future, I'd also like to test that it directs you to the right page when you click on it.)

I also changed a few things with the "creating a new session" tests to separate out the 'write to firebase' part from the 'redirect to a new page' part. They aren't the most isolated tests in the world, but it'll do.