Make the interact button scale up and down
We have to use vw and vh units, which is no fun, but other than that, it works in a pretty straight-forward manner. I didn't even need to turn off view encapsulation! :)

I haven't tested this on actual devices yet—I'll do that after I get the pollen counter to scale appropriately.