Make it so that each flower can only be visited once per nest cycle
While I was here, I refactored the Play Round Compontent and the Student Round Service  a bit.

The functionality seems to be all there, but I haven't put in any testing yet, so I'm not ready to open a pull request. (In particular, I'd very much like to look around for race conditions.)