Make sure to blur the inputs after you type stuff
To give the validation a chance to trigger.

Previously, I tried to solve this problem by changing when the
validation triggered, but that's not a very good solution, since we
actually don't mind the current validation behavior. We should try to
make Cypress cater to the validation behavior, rather than the other way
around.