Line up the pollen card and nest card
If we're going to add the flower names (and nest names), the layout
will be a lot cleaner if these cards have the same height.

To achieve this, we made the nest card one line instead of two lines,
and we changed how we were setting the font size. (Now, we use media
queries instead of calculating it from vw and vh.)