Use standard material breakpoints
Thanks to @floogulinc for pointing me in the direction of these. In
particular, we're now using 960px instead of 600px as the cutoff from
tablet-portrait to tablet-landscape.

These are just rough approximations, of course. For example, the
Samsung Galaxy Note 10+ in landscape mode falls in the "landscape
tablet" class, as does the 12-inch iPad Pro in portrait mode. (And then
there's the Microsoft Surface Duo, which is just kinda its own thing.)
The breakpoints given here are just meant to be suggestive! 🙃