Add Cypress tests for using the join code
Right now, we're just testing joining a session on the student side. We don't have tests for the page where the teacher makes a join code.

These tests are actually failing right now, because we haven't added a validation message for when you type in a malformed join code. We should really go do that.