Fix some flicker on the SessionLobbyComponent
Join codes were being read from the local cache before they had been written to the server, even if writing them to the server was going to fail.

Now, we handle this problem by filtering out values from the local cache, so that we only ever see values from the server. (We can tell which values are from the cache because they evaluate `firestore.FieldValue.serverTimestamp()` to null.)