Limit interaction to accepted_flowers
Continuing on with the idea that we want to provide individual tips for
students, I think we will need to a way to distinguish between the
reasons a flower can't be visited, so I made a 'knowsFlower' field.

I made a few comments throughout the code about tests for 'canVisits' will need to be redone.