Fix issue with height getting set to odd values when animation starts
Not quite sure what was happening but the height was sometimes getting set to 100% or some other value when the animation happened for the first time. Setting the initial height in the style versus the css class for the element seems to give anime.js the ability to know what said height is.