Get the GitHub SVG from the right place
We recently changed where the GitHub icon SVG lives. It used to be in
the `assets/` directory; now it's in a TS file. (We're statically
linking it in with the WebPack bundle.)

Unfortunately, one lingering reference to the `assets/` directory
didn't get changed. This commit fixes the issue.

(I also quickly searched through the codebase, and it looks like that
was the last reference to storing custom SVG icons in `assets/`.