Remove demo rounds, add two "polished" rounds
These are two of the the round templates we're actually planning on
using in production (although we'll probably make edits and revisions,
of course).

They're designed so that students will learn that different bees need
different flowers to survive; and that, if you only plant one or two
types of flowers, you're only serving the needs of a few bee species.

To this end, we've chosen five bee species with a diverse range of
flowers accepted--so that there's as little overlap as possible between
the diets of the different bees. That way, the students get a strong
sense that ecological diversity is important.

We've also tried to choose bees that have long active periods, so that
each student gets to play for quite a long time.