Add a moderately-monocultured round.
At Audrey's request; she thought that it would be good to show students
a gradient of monoculture/polyculture, so that they'd better be able to
draw conclusions about which is better. (They can see the trend line
easier if they have more data-points.)

I decided to make it so that the bees from round 1 that had an easier
time would now have a harder time, and vice versa. (To really make it
clear that a no matter how you restrict the set of flowers, it always
puts someone at a disadvantage.)