Remove customAssign function and change templates to just take an array of bee species
We are moving to the templates taking an array of bee species rather than an array of bees with weights. If you want to have some bee species weighted differently you can duplicate that species in the array. This allows us to use the much simpler and more reliable assignment algorithm.

Co-authored-by: Utkarsh Kumar <kumar375@morris.umn.edu>
Co-authored-by: KK Lamberty <lamberty@morris.umn.edu>