Rewrite some tests for TimerProgressSpinner
We've changed the way we think about elapsed time, so we need to update
the tests to reflect this.

For example, previously, if the round's end time was `x`, we considered
the round to be over as soon as the timer reaches `x`. Now that our
timer has finer resolution, we consider the round to be over after the
timer ticks all the way through `x` and reaches `x+1`.

As part of this change, we adjusted the way the TimerProgressSpinner
shows elapsed time. However, we hadn't updated the tests accordingly, so
they were failing. Now we've updated the tests.