Rewrite TimerService to do integer math internally
Previously, we were doing a lot of floating point divisions in the
timerState$ observable, which was causing a lot of rounding errors.

Now, as much as possible, we try to represent the timer's interal state
using integers, and to update the timer using integer math. We only
convert to a floating-point representation at the last possible moment.

This commit doesn't actually change TimerService's API, just how it
represents the time internally.