Refactor TimerService to track ms in state rather than ticks
This changes it so we track the elapsed milliseconds in the TimerState. This is accounting for the startTime so when initialized it is equal to startTime.time * tickSpeed (rounded to nearest multiple of baseTickSpeed). The end time in the state is also tracked in ms now with the extra tick at the end added in on initialization.