Stricten the field-guide e2e tests
Previously, when we opened a dialog, we tested that it contained the
common name somewhere. This isn't a perfect test, since dialogs usually
mention multiple species in passing. The test migh spuriously pass if
we opened a dialog for the wrong species, but the name of the correct
species happened to occur somewhere in the body.

So, in the revised version of the test, we check that the species name
occurs in the title of the dialong, not just anywhere in the body.

(All the tests are still passing. Also, this commit changes a type
annotation to be more specific.)