Fix the import of `three` in ArViewComponent
We were importing three from within threear's node_modules, which broke
when npm decided to rearrange where transitive dependencies were stored.

We can fix this problem by listing three as a direct dependency.
(This is a reasonable thing to do, since we're using it as a direct
dependency anyway.) I've specifically required three@^0.122.0, since
that's the same version that threear is using.

(Granted, we may have to take care to keep our versions of three and
threear in sync, but that's just how life is when you're using a library
directly and transitively ¯\_(ツ)_/¯)