Add (temporary) fix for field guide images in Webkit
This is a (hopefully temporary) fix for the issue we were having where images wouldn't display in the field guide dialog in Webkit (Safari). We now replace the images in the field guide dialog with the 512 square versions only on Webkit so we can set a default width/height for those.

Closes #257

Co-authored-by: Joe Walbran <walbr037@morris.umn.edu>